Fix: Alterar tipo do código de status do produto

+++ ProductGrpc/Models/Product.cs
@@ -7,7 +7,7 @@ namespace ProductGrpc.Models
         public string Name { get; set; }
         public string Description { get; set; }
         public float Price { get; set; }
-        public int StatusId {get;set; }
+        public ushort StatusId {get;set; }
         public virtual ProductStatus Status { get; set; }
     }
