+++ ProductGrpc/Models/ProductStatus.cs
@@ -1,6 +1,6 @@
 ﻿namespace ProductGrpc.Models
 {
-    public class ProductStatus : ModelBase<int>
+    public class ProductStatus : ModelBase<ushort>
     {
         public string Name { get; set; }
         public virtual IList<Product> Products { get; set; }
