+++ ProductGrpc/Services/Includes/IProductService.cs
@@ -6,7 +6,7 @@ namespace ProductGrpc.Services.Includes
     public interface IProductService : IBaseService<Product>
     {
         Task<IEnumerable<Product>> GetProducts(ProductFilter filter);
-        Task ChangeStatus(ProductIndexRequest request, int statusId);
+        Task ChangeStatus(ProductIndexRequest request, ushort statusId);
         Task ChangeDeletedState(ProductIndexRequest request, bool isDeleted);
     }
 }
